verze 0.0.3

+++ Bludiste/src/bludiste/Zed.java
@@ -7,12 +7,12 @@ import javax.swing.ImageIcon;
 
 public class Zed {
 	private String obr = "/zed.png";
-	private int x;
-	private int y;
+	private float x;
+	private float y;
 	private int width;
 	private int height;
 	private Image image;
-	public Zed(int x, int y) {
+	public Zed(float x, float y) {
 		ImageIcon obraz = new ImageIcon(this.getClass().getResource(obr));
 		image = obraz.getImage();
 		width = image.getWidth(null);
@@ -22,16 +22,16 @@ public class Zed {
 		
 	}
 	public int getX() {
-		return x;
+		return (int)x;
 	}
 	public int getY() {
-		return y;
+		return (int)y;
 	}
 	public Image getImage() {
 		return image;
 	}
 	public Rectangle getBounds(){
-		return new Rectangle(x, y, width, height);
+		return new Rectangle((int)x, (int)y, width, height);
 	}
 	public int getWidth(){
 		return width;
